Refactor: MessageCode -> CustomCode 객체 이름 변경
