+++ src/main/java/com/modong/backend/domain/form/FormController.java
@@ -2,7 +2,7 @@ package com.modong.backend.domain.form;
 
 import com.modong.backend.base.Dto.BaseResponse;
 import com.modong.backend.base.Dto.SavedId;
-import com.modong.backend.Enum.MessageCode;
+import com.modong.backend.Enum.CustomCode;
 import com.modong.backend.domain.form.dto.FormRequest;
 import com.modong.backend.domain.form.dto.FormResponse;
 import io.swagger.annotations.Api;
@@ -33,14 +33,14 @@ public class FormController {
   @Operation(summary = "페이지 생성", description = "지원서에 새로운 페이지를 저장한다. 질문 유형 QUESTION(1), SINGLE_SELECT_QUESTION(2), MULTI_SELECT_QUESTION(3)")
   public ResponseEntity createForm(@Valid @RequestBody FormRequest formRequest){
       SavedId savedId = new SavedId(formService.create(formRequest));
-      return ResponseEntity.ok(new BaseResponse(savedId, HttpStatus.CREATED.value(), MessageCode.SUCCESS_CREATE));
+      return ResponseEntity.ok(new BaseResponse(savedId, HttpStatus.CREATED.value(), CustomCode.SUCCESS_CREATE));
   }
   //페이지 수정
   @PutMapping("/form/{form_id}")
   @Operation(summary = "페이지 수정", description = "페이지의 ID를 이용해 수정한다. 질문 유형 QUESTION(1), SINGLE_SELECT_QUESTION(2), MULTI_SELECT_QUESTION(3)")
   public ResponseEntity updateForm(@Valid @PathVariable(name = "form_id") Long formId, @RequestBody FormRequest formRequest){
     SavedId savedId = new SavedId(formService.update(formId,formRequest));
-    return ResponseEntity.ok(new BaseResponse(savedId, HttpStatus.OK.value(), MessageCode.SUCCESS_UPDATE));
+    return ResponseEntity.ok(new BaseResponse(savedId, HttpStatus.OK.value(), CustomCode.SUCCESS_UPDATE));
   }
 
 
@@ -50,7 +50,7 @@ public class FormController {
   public ResponseEntity getFormById(@Valid @PathVariable(name = "form_id") Long formId){
 
     FormResponse form = formService.findById(formId);
-    return ResponseEntity.ok(new BaseResponse(form,HttpStatus.OK.value(),MessageCode.SUCCESS_GET));
+    return ResponseEntity.ok(new BaseResponse(form,HttpStatus.OK.value(), CustomCode.SUCCESS_GET));
   }
 
   //지원서 id로 해당되는 모든 페이지 조회
@@ -58,7 +58,7 @@ public class FormController {
   @Operation(summary = "지원서 ID 로 해당되는 모든 페이지 조회", description = "지원서 ID를 이용해 포함된 모든 페이지를 조회한다.")
   public ResponseEntity getFormsByApplicationId(@Valid @PathVariable(name = "application_id") Long applicationId){
     List<FormResponse> forms = formService.findAllByApplicationId(applicationId);
-    return ResponseEntity.ok(new BaseResponse(forms, HttpStatus.OK.value(), MessageCode.SUCCESS_GET_LIST));
+    return ResponseEntity.ok(new BaseResponse(forms, HttpStatus.OK.value(), CustomCode.SUCCESS_GET_LIST));
   }
 
   //지원서의 페이지 삭제
@@ -66,7 +66,7 @@ public class FormController {
   @Operation(summary = "페이지 삭제", description = "페이지의 ID를 이용해 삭제한다.")
   public ResponseEntity deleteFormById(@Valid @PathVariable(name = "form_id") Long formId){
     formService.deleteForm(formId);
-    return ResponseEntity.ok(new BaseResponse(HttpStatus.OK.value(),MessageCode.SUCCESS_DELETE));
+    return ResponseEntity.ok(new BaseResponse(HttpStatus.OK.value(), CustomCode.SUCCESS_DELETE));
   }
 
 }
