+++ src/main/java/com/modong/backend/domain/question/QuestionService.java
@@ -1,6 +1,6 @@
 package com.modong.backend.domain.question;
 
-import static com.modong.backend.Enum.MessageCode.ERROR_REQ_PARAM_ID;
+import static com.modong.backend.Enum.CustomCode.ERROR_REQ_PARAM_ID;
 
 import com.modong.backend.Enum.QuestionType;
 import com.modong.backend.domain.form.Form;
